+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://jeronlau.tk/semver/).
 ### Changed
  - `Executor::new()` now takes `impl Into<Arc<I>>` instead of `I`
  - `Executor::spawn()` no longer requires `Unpin` futures
+ - `Sleep` trait now requires `Send + Sync + 'static`
 
 ### Removed
  - **`std`** feature - to use pasts on no-std environments use the new